Honor Fast HLE (Mupen) choice for RSP plugin in mupen64plus-sa.
Prior to this, the script was always selecting Accurate HLE (CXD4)
regardless of the user's choice in Emulation Station.  It was always
falling through to the default case.  Now we properly catch the Fast HLE
choice.

Note that Accurate HLE (CXD4) is actually denoted in system.cfg with
"mle", not "cxd4", so that case was mislabeled for multiple reasons.

We still want CXD4 to be the default RSP plugin, which is why we only
fix the case labels and not the logic.